FishJoy/requirements.txt
asgiref==3.6.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-import-export==3.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
folium==0.14.0
fonttools==4.39.1
future==0.18.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
ratelim==0.1.6
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
